Added Feature: Able to support floating task add
Now whenever a floating task is to be added, we can call and write the
information into a text file called the floatingTask.txt. This allows us
to keep track of how many main events and floating task altogether. It
also allow us to display these 2 categories separately. Such support is
added in the storage.